--- a/packages/cli/package.json
+++ b/packages/cli/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@jimp/cli",
-  "version": "0.4.0",
+  "version": "0.5.0",
   "description": "jimp on the cli",
   "bin": {
     "jimp": "dist/index.js"
@@ -17,9 +17,9 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "@jimp/custom": "^0.4.0",
+    "@jimp/custom": "^0.5.0",
     "chalk": "^2.4.1",
-    "jimp": "^0.4.0",
+    "jimp": "^0.5.0",
     "log-symbols": "^2.2.0",
     "yargs": "^12.0.2"
   },