… packages/cli/test/images/action-print.png
… yarn.lock