# What's Changing and Why

Mix Before:
```
1704.093ms
1641.440ms
1684.027ms
1890.238ms
```

Mix After

```
265.268ms
258.583ms
264.966ms
263.853ms
```

the following also received the same perf boost

* tint
* shade
* xor 
* red
* green 
* blue

All other function still fall back to tinycolor and take around 5x as long

closes #270

## What else might be affected

Tests pass so I think we're good.

## Tasks

- [ ] Add tests
- [ ] Update Documentation
- [ ] Update `jimp.d.ts`
- [x] Add [SemVer](https://semver.org/) Label